make it ember-cli-sass 7.2.0 compatible
