+++ .travis.yml
@@ -28,6 +28,7 @@ env:
     - EMBER_TRY_SCENARIO=ember-lts-2.18
     - EMBER_TRY_SCENARIO=ember-release
     - EMBER_TRY_SCENARIO=ember-release-node-sass
+    - EMBER_TRY_SCENARIO=ember-release-with-old-ember-cli-sass
     - EMBER_TRY_SCENARIO=ember-beta
     - EMBER_TRY_SCENARIO=ember-canary
     - EMBER_TRY_SCENARIO=ember-default
