+++ index.js
@@ -1,6 +1,23 @@
 'use strict';
 const exportSass = require('./lib/exportSass');
 const path = require('path');
+const VersionChecker = require('ember-cli-version-checker');
+
+function getDefaultSass() {
+  try {
+    // eslint-disable-next-line node/no-extraneous-require, node/no-missing-require node/no-unpublished-require
+    return require('sass');
+  } catch (e) {
+    let error = new Error(
+      'Could not find the default SASS implementation. Run the default blueprint:\n' +
+      '   ember g ember-cli-sass\n' +
+      'Or install an implementation such as "node-sass" and add an implementation option. For example:\n' +
+      '   sassOptions: {implementation: require("node-sass")}');
+    error.type = 'Sass Plugin Error';
+
+    throw error;
+  }
+}
 
 module.exports = {
   name: 'ember-cli-sass-variables-export',
@@ -30,6 +47,23 @@ module.exports = {
       app.options.sassOptions.functions = {};
     }
 
-    Object.assign(app.options.sassOptions.functions, exportSass(path.join(__dirname, 'addon', 'utils'), app.options.sassOptions.implementation));
+    let checker = new VersionChecker(this);
+    let sass = checker.for('ember-cli-sass');
+
+    let implementation = app.options.sassOptions.implementation;
+
+    if (sass.lt('8.0.0')) {
+      // these ember-cli-sass versions used node-sass by default
+      // so we force its usage
+      // eslint-disable-next-line node/no-extraneous-require, node/no-missing-require
+      implementation = require('node-sass');
+    } else {
+      // on ember-cli-sass version8 and above
+      // we mimic the way ember-cli-sass tries to find the default sass implementation
+      // only runs when no implementation was passed in
+      implementation = implementation || getDefaultSass();
+    }
+
+    Object.assign(app.options.sassOptions.functions, exportSass(path.join(__dirname, 'addon', 'utils'), implementation));
   }
 };
