+++ package.json
@@ -40,7 +40,8 @@
     "postversion": "git push --follow-tags"
   },
   "dependencies": {
-    "ember-cli-babel": "^7.1.0"
+    "ember-cli-babel": "^7.1.0",
+    "ember-cli-version-checker": "^2.1.2"
   },
   "devDependencies": {
     "broccoli-asset-rev": "^3.0.0",
